DEVX-8 finalized alignment to VITE
