+++ .gitignore
@@ -1,3 +1,5 @@
 */.env
 */node_modules
+*/dist
+*/build
