+++ webauthn-cross-device/backend/desktop.js
@@ -25,7 +25,7 @@ async function getUserTokens(authCode) {
   const params = new URLSearchParams({
     grant_type: 'authorization_code',
     code: authCode,
-    client_id: process.env.TS_CLIENT_ID,
+    client_id: process.env.VITE_TS_CLIENT_ID,
     client_secret: process.env.TS_CLIENT_SECRET,
     redirect_uri: process.env.TS_REDIRECT_URI
   });
