+++ webauthn-cross-device/package.json
@@ -7,9 +7,6 @@
     "start": "vite --debug --host --port 8080"
   },
   "dependencies": {
-    "cookie-parser": "~1.4.6",
-    "debug": "~4.3.4",
-    "ejs": "~3.1.8",
     "express": "~4.18.2",
     "morgan": "~1.10.0",
     "node-fetch-commonjs": "3.2.4",
