+++ webauthn-cross-device/pages/config.js
@@ -1,0 +1,1 @@
+window.env = import.meta.env;
