+++ webauthn-cross-device/pages/desktop.html
@@ -3,8 +3,8 @@
     <title>Desktop Device </title>
     <link rel="stylesheet" href="/stylesheets/style.css" />
     <script type="text/javascript"  src="https://cdn.identity.security/authentication-sdk/2.3.0/TsAuthSdk.min.js"></script>
-    <script type="module">
-      const clientId = import.meta.env.VITE_TS_CLIENT_ID;
+    <script type="module" src="./config.js"></script>
+    <script>
       function updateElementText(id, text) {
         try {
           document.getElementById(id).textContent = text;
@@ -15,6 +15,7 @@
       // For more information see https://developer.transmitsecurity.com/guides/webauthn/cross_device_flows/#1-initialize-the-webauthn-sdk
       async function init() {
         try {
+          const clientId = window.env.VITE_TS_CLIENT_ID;
           updateElementText("clientId", clientId);
           await window.WebAuthnSdk.init(clientId, {serverPath: "https://webauthn.identity.security"})
           updateElementText('status', `SDK Init done`);
@@ -123,18 +124,14 @@
           updateElementText('status', `Error during token exchange for code ${authCode}: ${ex}`);
         }
       }
-
-      window.tsInit = init;
-      window.tsInitiateCrossDeviceFlow = initiateCrossDeviceFlow;
-
     </script>
   </head>
 
-  <body onload="tsInit()">
+  <body onload="init()">
     <h1>Desktop Device</h1>
     <span id="clientId"></span>
     <br><br>
-    <button id="start" onclick="tsInitiateCrossDeviceFlow()" disabled>Start</button>
+    <button id="start" onclick="initiateCrossDeviceFlow()" disabled>Start</button>
     <br><br>
     <img id="qr" src="" />
     <p id="url"></p>
