+++ webauthn-cross-device/sample.env
@@ -1,7 +1,7 @@
 # duplicate and rename this file to .env
 # Copy the base codespave URL, and add -8080.preview.app before the github.dev suffix. This is the base launch URL for configurations
 TS_REDIRECT_URI=<above-modified-url>/complete
-TS_CLIENT_ID=<your app alient id>
+VITE_TS_CLIENT_ID=<your app alient id> # VITE_ prefix allows exposing to client side pages as well as the backend
 TS_CLIENT_SECRET=<your app client secret>
 TS_APP_ID=<your app ID, will be used by the demo to point you to configuration pages>
